requirements.txt
Django==1.9.2
django-cors-headers==1.1.0
djangorestframework==3.3.2
wheel==0.26.0
